skills/material-template-filler/web/requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
python-docx>=0.8.11
